+++ library/src/lapack/roclapack_sygvj_hegvj.hpp
@@ -1,0 +1,118 @@
+/************************************************************************
+ * Derived from the BSD3-licensed
+ * LAPACK routine (version 3.7.0) --
+ *     Univ. of Tennessee, Univ. of California Berkeley,
+ *     Univ. of Colorado Denver and NAG Ltd..
+ *     December 2016
+ * Copyright (c) 2021-2022 Advanced Micro Devices, Inc.
+ * ***********************************************************************/
+
+#pragma once
+
+#include "rocblas.hpp"
+#include "roclapack_potrf.hpp"
+#include "roclapack_syev_heev.hpp"
+#include "roclapack_sygst_hegst.hpp"
+#include "rocsolver/rocsolver.h"
+
+template <bool BATCHED, bool STRIDED, typename T, typename S>
+void rocsolver_sygvj_hegvj_getMemorySize(const rocblas_eform itype,
+                                         const rocblas_evect evect,
+                                         const rocblas_fill uplo,
+                                         const rocblas_int n,
+                                         const rocblas_int batch_count,
+                                         size_t* size_scalars,
+                                         size_t* size_workArr)
+{
+    // if quick return no need of workspace
+    if(n == 0 || batch_count == 0)
+    {
+        *size_scalars = 0;
+        *size_workArr = 0;
+        return;
+    }
+
+    // size of scalars (constants)
+    *size_scalars = sizeof(T) * 3;
+
+    // size of array of pointers to workspace
+    if(BATCHED)
+        *size_workArr = sizeof(T*) * batch_count;
+    else
+        *size_workArr = 0;
+}
+
+template <typename T, typename S>
+rocblas_status rocsolver_sygvj_hegvj_argCheck(rocblas_handle handle,
+                                              const rocblas_eform itype,
+                                              const rocblas_evect evect,
+                                              const rocblas_fill uplo,
+                                              const rocblas_int n,
+                                              T A,
+                                              const rocblas_int lda,
+                                              T B,
+                                              const rocblas_int ldb,
+                                              S* residual,
+                                              const rocblas_int max_sweeps,
+                                              rocblas_int* n_sweeps,
+                                              S* W,
+                                              rocblas_int* info,
+                                              const rocblas_int batch_count = 1)
+{
+    // order is important for unit tests:
+
+    // 1. invalid/non-supported values
+    if(itype != rocblas_eform_ax && itype != rocblas_eform_abx && itype != rocblas_eform_bax)
+        return rocblas_status_invalid_value;
+    if(evect != rocblas_evect_none && evect != rocblas_evect_original)
+        return rocblas_status_invalid_value;
+    if(uplo != rocblas_fill_upper && uplo != rocblas_fill_lower)
+        return rocblas_status_invalid_value;
+
+    // 2. invalid size
+    if(n < 0 || lda < n || ldb < n || max_sweeps <= 0 || batch_count < 0)
+        return rocblas_status_invalid_size;
+
+    // skip pointer check if querying memory size
+    if(rocblas_is_device_memory_size_query(handle))
+        return rocblas_status_continue;
+
+    // 3. invalid pointers
+    if((n && !A) || (n && !B) || (n && !W) || (batch_count && !residual)
+       || (batch_count && !n_sweeps) || (batch_count && !info))
+        return rocblas_status_invalid_pointer;
+
+    return rocblas_status_continue;
+}
+
+template <bool BATCHED, bool STRIDED, typename T, typename S, typename U, bool COMPLEX = rocblas_is_complex<T>>
+rocblas_status rocsolver_sygvj_hegvj_template(rocblas_handle handle,
+                                              const rocblas_eform itype,
+                                              const rocblas_evect evect,
+                                              const rocblas_fill uplo,
+                                              const rocblas_int n,
+                                              U A,
+                                              const rocblas_int shiftA,
+                                              const rocblas_int lda,
+                                              const rocblas_stride strideA,
+                                              U B,
+                                              const rocblas_int shiftB,
+                                              const rocblas_int ldb,
+                                              const rocblas_stride strideB,
+                                              const S abstol,
+                                              S* residual,
+                                              const rocblas_int max_sweeps,
+                                              rocblas_int* n_sweeps,
+                                              S* W,
+                                              const rocblas_stride strideW,
+                                              rocblas_int* info,
+                                              const rocblas_int batch_count,
+                                              T* scalars,
+                                              T** workArr)
+{
+    ROCSOLVER_ENTER("sygvj_hegvj", "itype:", itype, "evect:", evect, "uplo:", uplo, "n:", n,
+                    "shiftA:", shiftA, "lda:", lda, "shiftB:", shiftB, "ldb:", ldb,
+                    "abstol:", abstol, "max_sweeps:", max_sweeps, "bc:", batch_count);
+
+    return rocblas_status_not_implemented;
+}